incubator/i3-2/requirements.txt
# Drift (i3-2) - The Reasoning Trader
# Dependencies

# Trading
alpaca-py>=0.10.0

# AI
anthropic>=0.34.0

# Utils
python-dotenv>=1.0.0
pytz>=2024.1
